Refactor into separate modules for constants, keys, apps, exceptions, and remote control
