+++ panasonic_viera/apps.py
@@ -1,0 +1,24 @@
+"""Enum class containing app product IDs for Panasonic Viera TV."""
+from enum import Enum
+
+
+class Apps(Enum):
+    """Contains several app product IDs."""
+
+    NETFLIX = "0010000200000001"
+    YOUTUBE = "0070000200180001"
+    SHOUTCAST = "0070000400000001"
+    CALENDAR = "0387878700150020"
+    BROWSER = "0077777700160002"
+    AMAZONPRIME = "0010000100180001"
+    IPLAYER = "0020000A00000010"
+    BBCIPLAYER = "0020000A00000010"
+    ITV = "0387878700000124"
+    ALL4 = "0387878700000125"
+    DEMAND5 = "0020009300000001"
+    RECORDEDTV = "0387878700000013"
+    MULTIWINDOW = "0387878700000050"
+    BBCNEWS = "0020000A00000006"
+    BBCSPORT = "0020000A00000007"
+    WEATHER = "0070000C00000001"
+    DEVELOPER = "0077777777777778"
