+++ panasonic_viera/keys.py
@@ -1,0 +1,90 @@
+"""Enum class containing all known keys for Panasonic Viera TV remote control."""
+from enum import Enum
+
+
+class Keys(Enum):
+    """Contains all known keys."""
+
+    THIRTY_SECOND_SKIP = "NRC_30S_SKIP-ONOFF"
+    TOGGLE_3D = "NRC_3D-ONOFF"
+    APPS = "NRC_APPS-ONOFF"
+    ASPECT = "NRC_ASPECT-ONOFF"
+    BACK = "NRC_RETURN-ONOFF"
+    BLUE = "NRC_BLUE-ONOFF"
+    CANCEL = "NRC_CANCEL-ONOFF"
+    CC = "NRC_CC-ONOFF"
+    CHAT_MODE = "NRC_CHAT_MODE-ONOFF"
+    CH_DOWN = "NRC_CH_DOWN-ONOFF"
+    INPUT_KEY = "NRC_CHG_INPUT-ONOFF"
+    NETWORK = "NRC_CHG_NETWORK-ONOFF"
+    CH_UP = "NRC_CH_UP-ONOFF"
+    NUM_0 = "NRC_D0-ONOFF"
+    NUM_1 = "NRC_D1-ONOFF"
+    NUM_2 = "NRC_D2-ONOFF"
+    NUM_3 = "NRC_D3-ONOFF"
+    NUM_4 = "NRC_D4-ONOFF"
+    NUM_5 = "NRC_D5-ONOFF"
+    NUM_6 = "NRC_D6-ONOFF"
+    NUM_7 = "NRC_D7-ONOFF"
+    NUM_8 = "NRC_D8-ONOFF"
+    NUM_9 = "NRC_D9-ONOFF"
+    DIGA_CONTROL = "NRC_DIGA_CTL-ONOFF"
+    DISPLAY = "NRC_DISP_MODE-ONOFF"
+    DOWN = "NRC_DOWN-ONOFF"
+    ENTER = "NRC_ENTER-ONOFF"
+    EPG = "NRC_EPG-ONOFF"
+    EXIT = "NRC_CANCEL-ONOFF"
+    EZ_SYNC = "NRC_EZ_SYNC-ONOFF"
+    FAVORITE = "NRC_FAVORITE-ONOFF"
+    FAST_FORWARD = "NRC_FF-ONOFF"
+    GAME = "NRC_GAME-ONOFF"
+    GREEN = "NRC_GREEN-ONOFF"
+    GUIDE = "NRC_GUIDE-ONOFF"
+    HOLD = "NRC_HOLD-ONOFF"
+    HOME = "NRC_HOME-ONOFF"
+    INDEX = "NRC_INDEX-ONOFF"
+    INFO = "NRC_INFO-ONOFF"
+    CONNECT = "NRC_INTERNET-ONOFF"
+    LEFT = "NRC_LEFT-ONOFF"
+    MENU = "NRC_MENU-ONOFF"
+    MPX = "NRC_MPX-ONOFF"
+    MUTE = "NRC_MUTE-ONOFF"
+    NET_BS = "NRC_NET_BS-ONOFF"
+    NET_CS = "NRC_NET_CS-ONOFF"
+    NET_TD = "NRC_NET_TD-ONOFF"
+    OFF_TIMER = "NRC_OFFTIMER-ONOFF"
+    PAUSE = "NRC_PAUSE-ONOFF"
+    PICTAI = "NRC_PICTAI-ONOFF"
+    PLAY = "NRC_PLAY-ONOFF"
+    P_NR = "NRC_P_NR-ONOFF"
+    POWER = "NRC_POWER-ONOFF"
+    PROGRAM = "NRC_PROG-ONOFF"
+    RECORD = "NRC_REC-ONOFF"
+    RED = "NRC_RED-ONOFF"
+    RETURN_KEY = "NRC_RETURN-ONOFF"
+    REWIND = "NRC_REW-ONOFF"
+    RIGHT = "NRC_RIGHT-ONOFF"
+    R_SCREEN = "NRC_R_SCREEN-ONOFF"
+    LAST_VIEW = "NRC_R_TUNE-ONOFF"
+    SAP = "NRC_SAP-ONOFF"
+    TOGGLE_SD_CARD = "NRC_SD_CARD-ONOFF"
+    SKIP_NEXT = "NRC_SKIP_NEXT-ONOFF"
+    SKIP_PREV = "NRC_SKIP_PREV-ONOFF"
+    SPLIT = "NRC_SPLIT-ONOFF"
+    STOP = "NRC_STOP-ONOFF"
+    SUBTITLES = "NRC_STTL-ONOFF"
+    OPTION = "NRC_SUBMENU-ONOFF"
+    SURROUND = "NRC_SURROUND-ONOFF"
+    SWAP = "NRC_SWAP-ONOFF"
+    TEXT = "NRC_TEXT-ONOFF"
+    TV = "NRC_TV-ONOFF"
+    UP = "NRC_UP-ONOFF"
+    LINK = "NRC_VIERA_LINK-ONOFF"
+    VOLUME_DOWN = "NRC_VOLDOWN-ONOFF"
+    VOLUME_UP = "NRC_VOLUP-ONOFF"
+    VTOOLS = "NRC_VTOOLS-ONOFF"
+    YELLOW = "NRC_YELLOW-ONOFF"
+    HMDI_1 = "NRC_HDMI1-ONOFF"
+    HMDI_2 = "NRC_HDMI2-ONOFF"
+    HMDI_3 = "NRC_HDMI3-ONOFF"
+    HMDI_4 = "NRC_HDMI4-ONOFF"
