+++ pyproject.bak.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "panasonic-viera"
+version = "0.4.2"
+description = "Library to control Panasonic Viera TVs"
+authors = ["Florian Holzapfel <flo.holzapfel@gmail.com>"]
+license = "MIT"
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.8"
+aiohttp = "^3.9.5"
+pycryptodome = "^3.20.0"
+xmltodict = "^0.13.0"
+
+[tool.poetry.group.dev.dependencies]
+ruff = "^0.4.1"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
